requirements.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pycocotools==2.0.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.4.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchaudio==2.7.0+cpu
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
